Move properties out of condition.go
